CMake fixes for Windows. Warning fixes for Windows. Created CMake Find... modules for SQLite, MySQL, and WiiYourself. Fixed compiler warnings for Windows.

+++ time/StdTime.hpp
@@ -52,7 +52,7 @@ inline
 double 
 StdTime::MaxSeconds( ) const
 {
-    return std::numeric_limits< std::time_t >::max();
+    return (double)std::numeric_limits< std::time_t >::max();
 }
